+++ mod.testing.ts
@@ -1,3 +1,4 @@
 // Copyright 2023 Im-Beast. MIT license.
 export * from "./src/testing/input.ts";
 export * from "./src/testing/snapshot.ts";
+export * from "./src/testing/app.ts";